Add an option to turn off script name caching to improve preformance.
Script names are cached to allow features like compile-time valididy
checking, or the ability to right click on a script name to bring up
the corresponding creature or area container. However, this comes with
a performance cost, which may not be worth it under some circumstances.
Add an option to the BrowserMenuBar to disable this behaviour and make
setupScriptNames in the BCS compiler conditional on this option.