Bugfix with resources loaded from BCS
If you had, for example, a creature and an item of the same resref,
all triggers and actions with a S:ResRef* parameter would load the
creature, even if the item should have been loaded (e.g., if it was
the HasItem trigger). This was so for
CRE > ITM > ARE > 2DA > BCS > MVE > SPL > DLG ...
This problem could also shadow missing resources, if there existed a
matching resref of higher priority but of wrong type.