Add new feature: Perform WeiDU changelog
Performs a WeiDU changelog on a selected game resource and presents the
result as table of changelog entries. Option is available from File menu
or resource tree context menu.
Intermediate resources associated with the individual changelog entries
can be opened in NI.
A Weidu binary is needed for the operation, which can be defined in the
Preferences, or is autodetected if available in the game directory.